Moved distributedLockHelper to integrations project
