+++ src/Altinn.Correspondence.Application/MalwareScanResult/MalwareScanResultHandler.cs
@@ -1,6 +1,6 @@
 ﻿using Altinn.Correspondence.Application.Helpers;
 using Altinn.Correspondence.Application.MalwareScanResult.Models;
-using Altinn.Correspondence.Common.Caching;
+using Altinn.Correspondence.Integrations.Redlock;
 using Altinn.Correspondence.Core.Models.Entities;
 using Altinn.Correspondence.Core.Models.Enums;
 using Altinn.Correspondence.Core.Repositories;
