+++ src/Altinn.Correspondence.Common/Altinn.Correspondence.Common.csproj
@@ -8,8 +8,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="9.0.2" />
-    <PackageReference Include="Microsoft.Extensions.Caching.StackExchangeRedis" Version="9.0.2" />
-    <PackageReference Include="RedLock.net" Version="2.3.2" />
     <PackageReference Include="Serilog" Version="4.2.0" />
   </ItemGroup>
