+++ src/Altinn.Correspondence.Integrations/DependencyInjection.cs
@@ -13,6 +13,7 @@ using Altinn.Correspondence.Integrations.Altinn.Register;
 using Altinn.Correspondence.Integrations.Altinn.ResourceRegistry;
 using Altinn.Correspondence.Integrations.Altinn.Storage;
 using Altinn.Correspondence.Integrations.Azure;
+using Altinn.Correspondence.Integrations.Redlock;
 using Altinn.Correspondence.Integrations.Dialogporten;
 using Altinn.Correspondence.Integrations.Slack;
 using Microsoft.Extensions.Configuration;
@@ -65,6 +66,7 @@ public static class DependencyInjection
         }
         
         services.AddSingleton<SlackSettings>();
+        services.AddSingleton<DistributedLockHelper>();
     }
 
     public static void RegisterAltinnHttpClient<TClient, TImplementation>(this IServiceCollection services, MaskinportenSettings maskinportenSettings, AltinnOptions altinnOptions)
