+++ src/Altinn.Correspondence.Integrations/Redlock/DistributedLockHelper.cs
@@ -1,11 +1,11 @@
-using Microsoft.Extensions.Caching.StackExchangeRedis;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using RedLockNet.SERedis;
 using RedLockNet.SERedis.Configuration;
 using StackExchange.Redis;
-using System.Reflection;
+using Altinn.Correspondence.Core.Options;
 
-namespace Altinn.Correspondence.Common.Caching
+namespace Altinn.Correspondence.Integrations.Redlock
 {
     public class DistributedLockHelper
     {
@@ -15,28 +15,32 @@ namespace Altinn.Correspondence.Common.Caching
         public const int DefaultRetryDelayMs = 500;
         private const string LockKeyPrefix = "lock:";
 
-        public DistributedLockHelper(RedisCache redisCache, ILogger<DistributedLockHelper> logger)
+        public DistributedLockHelper(IOptions<GeneralSettings> generalSettings, ILogger<DistributedLockHelper> logger)
         {
-            ArgumentNullException.ThrowIfNull(redisCache);
             _logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
-            var propertyInfo = typeof(RedisCache).GetProperty("ConnectionMultiplexer", 
-                BindingFlags.NonPublic | BindingFlags.Instance) 
-                ?? throw new InvalidOperationException("ConnectionMultiplexer property not found on RedisCache");
+            var redisConnectionString = generalSettings?.Value?.RedisConnectionString;
+            if (string.IsNullOrEmpty(redisConnectionString))
+            {
+                redisConnectionString = Environment.GetEnvironmentVariable("GeneralSettings__RedisConnectionString");
+            }
+            
+            if (string.IsNullOrEmpty(redisConnectionString))
+            {
+                throw new InvalidOperationException("Redis connection string not found. Please add 'RedisConnectionString' to GeneralSettings in your configuration or set the 'GeneralSettings__RedisConnectionString' environment variable.");
+            }
 
-            var connectionMultiplexer = propertyInfo.GetValue(redisCache) as ConnectionMultiplexer 
-                ?? throw new InvalidOperationException("Failed to get ConnectionMultiplexer from RedisCache");
-
+            var connectionMultiplexer = ConnectionMultiplexer.Connect(redisConnectionString);
             var multiplexers = new List<RedLockMultiplexer> { new RedLockMultiplexer(connectionMultiplexer) };
             _lockFactory = RedLockFactory.Create(multiplexers);
         }
 
         /// <summary>
-        /// Acquires a distributed lock and executes the action when lock is aquired.
+        /// Acquires a distributed lock and executes the action if successful.
         /// The lock is released automatically when the action is completed.
         /// </summary>
         /// <param name="lockKey">The unique key for the lock</param>
-        /// <param name="action">The action to execute when lock is acquired</param>
+        /// <param name="action">The action to execute if lock is acquired</param>
         /// <param name="retryCount">Number of retries if lock acquisition fails (default: 3)</param>
         /// <param name="retryDelayMs">Delay between retries in milliseconds (default: 500)</param>
         /// <param name="cancellationToken">Cancellation token</param>
